00-requirements.txt
google-generativeai>=0.3
streamlit>=1.34